Declare hole in 32u4 analog pin numbering.
Just having a macro to tell whether an ADC pin is there is not sufficient,
the TOTAL_ANALOG_PINS could be use to make up a non-trivial iterator but
for backward compatibility we do not want to reuse it anyway.  So we provide
a LAST_ANALOG_PIN for implementation of a trivial iterator.