Let SerialLCD wait for UART_READY before sending INIT_ACK.
Without this, on 32u4 the INIT_ACK is sent before we can be sure the
UART_READY was even sent, and the simple Blink example does not work.

This patch enforces the documented init protocol, which has the added
bonus of allowing to get rid of ugly delay() calls, which were probably
introduced to kludge around timings issues.